nlwSpacetime - OAuth e rota para memories, feita no Web e Mobile

+++ nlwSpacetime-12/backend/src/routes/memories.ts
@@ -5,9 +5,18 @@ import { z } from 'zod'
 
 export async function memoriesRoutes(app: FastifyInstance) {
 
+  //Verifica se o usuário está logado.
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()   
+  })
+
   //Listar
-  app.get('/memories', async () => {
+  app.get('/memories', async (request) => {
+
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -23,7 +32,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
   })
 
   //Buscar
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     }) 
@@ -35,6 +44,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       }
     })
+
+    if (!memory.isPublic &&  memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     return memory
   })
@@ -54,7 +67,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: 'ce310683-519a-4104-97ca-5eb4396206a0',
+        userId: request.user.sub,
       }
     })
 
@@ -63,7 +76,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
   })
 
   //Atualizar
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     }) 
@@ -79,7 +92,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -94,14 +117,24 @@ export async function memoriesRoutes(app: FastifyInstance) {
   })
 
   //Delete
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     }) 
 
     const { id } = paramsSchema.parse(request.params)
 
-    const memory = await prisma.memory.delete({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.delete({
       where: {
         id,
       }
